Added blocking for Staging API

+++ main.py
@@ -51,6 +51,8 @@ def gitstats():
 # Core API to Add Data to Firestore + Push messages via Pushbullet
 @app.route('/api', methods=['POST'])  # GET requests will be blocked
 def add():
+    if(my_directory == "/home/stagingapi/mysite"):
+        return("Blocked. This is is not available on the staging API.", 401)
     # https://scotch.io/bar-talk/processing-incoming-request-data-in-flask
     req_data = request.get_json()
     Page = req_data['Page']
@@ -77,6 +79,8 @@ def add():
 # Route to Delete All Pushbullet Notifications. # Route to Shut Down API. Uses 256-bit key encryption.
 @app.route('/pbdel', methods=['GET'])
 def pbdelete():
+    if(my_directory == "/home/stagingapi/mysite"):
+        return("Blocked. This is is not available on the staging API.", 401)
     AuthCode = request.args.get('auth')
     if(AuthCode == Auth_Token):
         return(pushbullet.delete())
@@ -86,6 +90,8 @@ def pbdelete():
 # Route for SMS. Uses Twilio API
 @app.route("/sms", methods=["POST"])
 def sms_reply():
+    if(my_directory == "/home/stagingapi/mysite"):
+        return("Blocked. This is is not available on the staging API.", 401)
     message_content = request.values.get('Body', None)
     contact = request.values.get('From', None)
     try:
